% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{scale_rm}
\alias{scale_rm}
\title{Standardises matrices with multiple measurements per individual.}
\usage{
scale_rm(mat, origin = NULL, centre = F)
}
\arguments{
\item{mat}{A matrix.}

\item{origin}{Measurement of reference for stardadisation.}

\item{centre}{Logical. TRUE to centre data. Default is FALSE.}
}
\description{
This function stardadises matrices with multiple measurements
w.r.t. a chosen origin.
}
