% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tosccamm_folds.R
\name{tosccamm}
\alias{tosccamm}
\title{Computes TOSCCA-MM}
\usage{
tosccamm(
  A,
  B,
  nonzero_a,
  nonzero_b,
  folds = 1,
  parallel_logic = FALSE,
  silent = FALSE,
  ATest_res = NULL,
  BTest_res = NULL,
  model = "lme",
  lmeformula = " ~ -1 + time + (1|id)",
  arformula = c(1, 0, 0)
)
}
\arguments{
\item{A}{A matrix.}

\item{B}{A matrix.}

\item{nonzero_a}{Integer. Threshold parameter for A.}

\item{nonzero_b}{Integer. Threshold parameter for B.}

\item{folds}{Integer. Indicates number of folds to perform.}

\item{parallel_logic}{Logical. TRUE to parallelise folds. Default is FALSE.}

\item{silent}{Logical. TRUE to keep silent output messages. Default is FALSE.}

\item{ATest_res}{NULL. Keep NULL.}

\item{BTest_res}{NULL. Keep NULL.}

\item{model}{Character. c("lme", "ar"). Model to fit longitudinal latent space.}

\item{lmeformula}{Character. LME formula. Default is " ~ -1 + time + (1|id)".}

\item{arformula}{Numeric vector. Choice of ARIMA. Default is c(1,0,0).}
}
\value{
Canonical vectors for k components.
}
\description{
This function estimates sparse canonical vectors for matrices with multiple measurements
and the trajectories of the latent variables.
}
