% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{myHeatmap}
\alias{myHeatmap}
\title{Plot heatmap of cv w.r.t. the penalty parameter perfotmance.}
\usage{
myHeatmap(
  mat,
  palette = "rocket",
  xlab = "",
  ylab = "",
  show_axes = TRUE,
  show_labels = TRUE,
  K = 1
)
}
\arguments{
\item{mat}{A matrix.}

\item{palette}{Character. Name of a palette for the heatmap. Default is "Teal".}

\item{xlab}{Character. Label for x axis.}

\item{ylab}{Character. Label for y axis.}

\item{show_axes}{Logic. Default is True.}

\item{show_labels}{Logic. Default is True.}

\item{K}{Numeric. Number of components.}
}
\description{
This function plots cca for different thresholds
}
