% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{myHeatmap}
\alias{myHeatmap}
\title{Plot heatmap of cv w.r.t. the penalty parameter perfotmance.}
\usage{
myHeatmap(mat, palette = "Teal", coln = 12, xlab = "", ylab = "", axes = FALSE)
}
\arguments{
\item{mat}{A matrix.}

\item{palette}{Character. Name of a palette for the heatmap. Default is "Teal".}
}
\description{
This function stardadises matrices with multiple measurements
w.r.t. a chosen origin.
}
