% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{plt.selstab}
\alias{plt.selstab}
\title{Plot slection stability for penalty parameter performance.}
\usage{
plt.selstab(
  object,
  X,
  Y,
  palette_values = tosccamm:::mpalette,
  blue = T,
  mm = T,
  k = 1
)
}
\arguments{
\item{X}{nxp matrix. Observation matrix.}

\item{Y}{A nxq matrix. Observation matrix.}

\item{palette_values}{Character. Name of a palette for the heatmap. Default is "Teal".}

\item{mm}{Logic. Indicates whether there are multiple measurements or not. Default is True.}

\item{mat}{A matrix of canonical correlation for different sparsity levels.}
}
\description{
This function plots cv for different thresholds
}
