% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tosccam_permut.R
\name{toscamm.perm}
\alias{toscamm.perm}
\title{Computes permutatied cc fot TOSCCA-MM}
\usage{
toscamm.perm(
  A,
  B,
  nonzero_a,
  nonzero_b,
  K,
  alpha_init = c("eigen", "random", "uniform"),
  folds = 1,
  toPlot = FALSE,
  draws = 1000,
  cancor,
  bootCCA = NULL,
  silent = TRUE,
  parallel_logic = TRUE,
  nuisanceVar = 0,
  testStatType = "CC",
  model = "lme",
  lmeforemula = " ~ 0 + poly(time,3) + (1|id)",
  arformula = NULL
)
}
\arguments{
\item{A}{A matrix.}

\item{B}{A matrix.}

\item{nonzero_a}{Integer. Threshold parameter for A.}

\item{nonzero_b}{Integer. Threshold parameter for B.}

\item{K}{Interger. Numner of components.}

\item{alpha_init}{Character. Type of canonical vector initialisation c("uniform", "random", "eigen")}

\item{folds}{Integer. Indicates number of folds to perform.}

\item{toPlot}{Logical. Indicates if plots will be produced. Default is False.}

\item{draws}{Integer. Number of draws in permutation.}

\item{cancor}{Numeric vector of length K with estimated canonical correlations.}

\item{bootCCA}{deprecated.}

\item{silent}{Logical. TRUE to keep silent output messages. Default is FALSE.}

\item{parallel_logic}{Logical. TRUE to parallelise folds. Default is FALSE.}

\item{nuisanceVar}{Numeric. Number of nuisance variables.}

\item{testStatType}{Character. Choice of test-statistic c("CC", "Wilks", "Roy"),}

\item{model}{Character. c("lme", "ar"). Model to fit longitudinal latent space.}

\item{lmeforemula}{Character. LME formula. Default is " ~ -1 + time + (1|id)".}

\item{arformula}{Numeric vector. Choice of ARIMA. Default is c(1,0,0).}
}
\value{
Permutated canonical correlation for ell K and p-values.
}
\description{
This function estimates sparse canonical vectors for permuted matrices with multiple measurements.
}
